+++ src/main/java/de/unihalle/informatik/rhizoTrak/addon/RhizoLayerInfo.java
@@ -47,14 +47,36 @@
 
 package de.unihalle.informatik.rhizoTrak.addon;
 
+import java.awt.Point;
+import java.awt.geom.Point2D;
 import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.Vector;
 
+import javax.swing.tree.DefaultMutableTreeNode;
+import javax.swing.tree.DefaultTreeModel;
+import javax.swing.tree.TreePath;
+
+import de.unihalle.informatik.rhizoTrak.Project;
+import de.unihalle.informatik.rhizoTrak.display.Display;
+import de.unihalle.informatik.rhizoTrak.display.Displayable;
 import de.unihalle.informatik.rhizoTrak.display.Layer;
 import de.unihalle.informatik.rhizoTrak.display.Patch;
+import de.unihalle.informatik.rhizoTrak.display.Polyline;
 import de.unihalle.informatik.rhizoTrak.display.Treeline;
+import de.unihalle.informatik.rhizoTrak.tree.DNDTree;
+import de.unihalle.informatik.rhizoTrak.tree.ProjectThing;
+import de.unihalle.informatik.rhizoTrak.tree.ProjectTree;
+import de.unihalle.informatik.rhizoTrak.tree.TemplateThing;
+import de.unihalle.informatik.rhizoTrak.tree.TemplateTree;
+import de.unihalle.informatik.rhizoTrak.utils.Utils;
 import de.unihalle.informatik.rhizoTrak.xsd.rsml.RootType;
 import de.unihalle.informatik.rhizoTrak.xsd.rsml.Rsml;
 import de.unihalle.informatik.rhizoTrak.xsd.rsml.Rsml.Scene;
+import ij.gui.PolygonRoi;
+import ij.gui.Roi;
 
 /** Hold all information associated with a layer, especially including information related to RSML import/export
  * 
@@ -62,6 +84,11 @@ import de.unihalle.informatik.rhizoTrak.xsd.rsml.Rsml.Scene;
  *
  */
 public class RhizoLayerInfo {
+	
+	/**
+	 * Reference to the project to which the layer associated with this info belongs to.
+	 */
+	private Project project;
 	
 	/**
 	 * the layer the rsml has been imported into
@@ -81,12 +108,12 @@ public class RhizoLayerInfo {
 	/**
 	 * ROI associated with the layer, maybe null.
 	 */
-	RhizoROI roi = null;
+	private RhizoROI roi = null;
 	
 	/**
-	 * Gravitational direction in image of layer.
+	 * Gravitational direction in image of layer, maybe null.
 	 */
-	RhizoGravitationalDirection gravDir = null;
+	private RhizoGravitationalDirection gravDir = null;
 	
 	/**
 	 * map the treelines generated to the source RSML (top level) roots
@@ -98,7 +125,8 @@ public class RhizoLayerInfo {
 	 */
 	HashMap<RootType,Scene.Plant> rootPlantMap = new HashMap<RootType,Scene.Plant>();	
 	
-	public RhizoLayerInfo( Layer layer, Rsml rsml) {
+	public RhizoLayerInfo( Project p, Layer layer, Rsml rsml) {
+		this.project = p;
 		this.layer = layer;
 		this.rsml = rsml;
 		updateImageHash();
@@ -120,30 +148,6 @@ public class RhizoLayerInfo {
 		return rootPlantMap.get( root);
 	}
 
-	/**
-	 * Set ROI for this layer.
-	 * @param r	ROI object.
-	 */
-	public void setROI(RhizoROI r) {
-		this.roi = r;
-	}
-	 
-	/**
-	 * Get ROI for layer.
-	 * @return	ROI object.
-	 */
-	public RhizoROI getROI() {
-		return this.roi;
-	}
-	
-	public void setGravitationalDirection(RhizoGravitationalDirection gd) {
-		this.gravDir = gd;
-	}
-	
-	public RhizoGravitationalDirection getGravitationalDirection() {
-		return this.gravDir;
-	}
-	
 	/**
 	 * @return the rsml
 	 */
@@ -182,4 +186,256 @@ public class RhizoLayerInfo {
 		this.imageHash = RhizoUtils.calculateSHA256(patch.getFilePath());
     	return true;
 	}
+	
+	/*
+	 * ROI related stuff.
+	 */
+	
+	/**
+	 * Set the ROI for this layer from ImageJ Roi.
+	 * @param roi	 ImageJ ROI from which to initialize the internal ROI object.
+	 */
+	public void setROI(Roi roi) {
+		
+    // make new closed polyline as ROI and add to current layer
+    try {
+    	this.addROI(roi);
+    } catch (Exception e) {
+    		e.printStackTrace();
+        Utils.showMessage("Cannot create ROI from ImageJ Roi, adding ROI failed!");
+        return;
+    }
+
+	}
+			
+	/**
+	 * Set ROI for this layer from point list, e.g., read from RSML files.
+	 * @param points	List of points of ROI polygon.
+	 */
+	public void setROI(List<Point2D.Double> points) {
+
+		// create ImageJ Roi
+    float[] xPoints = new float[points.size()];
+		float[] yPoints = new float[points.size()];
+		for (int i=0; i<points.size(); ++i) {
+			xPoints[i] = (float)points.get(i).x;
+			yPoints[i] = (float)points.get(i).y;
+		}
+		PolygonRoi pRoi = new PolygonRoi(xPoints, yPoints, Roi.POLYGON);
+		
+    // make new closed polyline as ROI and add to current layer
+    try {
+    	this.addROI(pRoi);
+    } catch (Exception e) {
+    		e.printStackTrace();
+        Utils.showMessage("Cannot create ROI from points, adding ROI failed!");
+        return;
+    }
+	}
+	 
+  /**
+   * Add ROI to project.
+   * @param roi	Roi object to set as ROI for this layer.
+   */
+  private void addROI(Roi roi) {
+  	
+		// delete old ROI if there is one
+		this.clearROI();
+		
+		// get the parent thing in project tree where to add the ROI as a child
+		ProjectThing roiParentThing = this.getParentThing("roi");
+
+    ProjectThing pt = roiParentThing.createChild("polyline");
+    Polyline newPolyline = (Polyline) pt.getObject();
+
+    int n;
+    for (n = 0; n < roi.getPolygon().npoints; n++) {
+        newPolyline.insertPoint(n, roi.getPolygon().xpoints[n], roi.getPolygon().ypoints[n], 
+        		layer.getId());
+    }
+    newPolyline.insertPoint(n, roi.getPolygon().xpoints[0], roi.getPolygon().ypoints[0], 
+    		layer.getId());
+
+    // add new polyline to the project tree
+    ProjectTree currentTree = this.project.getProjectTree();
+    DefaultMutableTreeNode parentNode = DNDTree.findNode(roiParentThing, currentTree);
+    DefaultMutableTreeNode node = new DefaultMutableTreeNode(pt);
+    ((DefaultTreeModel) currentTree.getModel()).insertNodeInto(node, parentNode, 
+    		parentNode.getChildCount());
+    
+    newPolyline.setVisible(true, true);
+    pt.setVisible(true);
+    // BM: not sure why, but this seems to be essential here to ensure that the ROI is visible
+    newPolyline.repaint(true, layer);
+
+    // store new ROI object
+    this.roi = new RhizoROI(roi, newPolyline);
+  }
+  
+	/**
+	 * Get ROI for layer.
+	 * @return	ROI object.
+	 */
+	public RhizoROI getROI() {
+		return this.roi;
+	}
+	
+	/**
+	 * Get points of ROI.
+	 * @return	List of ROI points.
+	 */
+	public List<Point2D.Double> getROIPoints() {
+		Vector<Point2D.Double> points = new Vector<Point2D.Double>();
+		for (Point p : this.roi.getRoi())
+			points.add(new Point2D.Double(p.x, p.y));
+		return points;
+	}
+
+	/**
+	 * Delete the ROI from info object and also from the project tree.
+	 */
+	public void clearROI() {
+		
+    // find roi child of rootstack
+    ProjectThing roiProjectThing = this.getParentThing("roi");
+    if (roiProjectThing == null) {
+    	return;
+    }
+
+		Set<Displayable> deleteSet = new HashSet<Displayable>();
+
+  	for (ProjectThing ptc : roiProjectThing.findChildrenOfTypeR(Polyline.class)) {
+  		Polyline pl = (Polyline) ptc.getObject();
+  		if (layer == null || pl.getFirstLayer().getId() == layer.getId()) {
+  			deleteSet.add(pl);
+  		}
+  	}
+
+  	if (!deleteSet.isEmpty()) {
+  		this.project.removeAll(deleteSet);
+  	}
+  	
+  	// delete ROI object itself
+		this.roi = null;
+	}
+	
+	/*
+	 * Gravitational direction related stuff.
+	 */
+	
+	/**
+	 * Set gravitational direction and add a proper segment to project.
+	 * @param direction	Direction of gravitation in layer.
+	 */
+	public void setGravitationalDirection(double direction) {
+		
+		// delete old ROI if there is one
+		this.clearGravitationalDirection();
+		
+		// get the parent thing in project tree where to add the ROI as a child
+		ProjectThing dirParentThing = this.getParentThing("gravitationaldirection");
+		
+		// add new polyline to project
+		Polyline newPolyline = null;
+
+		// make new polyline for gravitational direction in current layer
+		ProjectThing pt = dirParentThing.createChild("polyline");
+		newPolyline = (Polyline) pt.getObject();
+		newPolyline.insertPoint(0, 10, 10, this.layer.getId());
+		newPolyline.insertPoint(1, (int)(10 + Math.sin(Math.toRadians(direction))*100),
+				(int)(10 + Math.cos(Math.toRadians(direction))*100), this.layer.getId());
+
+		// add new polyline to the project tree
+		ProjectTree currentTree = this.project.getProjectTree();
+		DefaultMutableTreeNode parentNode = DNDTree.findNode(dirParentThing, currentTree);
+		DefaultMutableTreeNode node = new DefaultMutableTreeNode(pt);
+		((DefaultTreeModel) currentTree.getModel()).insertNodeInto(node, parentNode, 
+				parentNode.getChildCount());
+
+		newPolyline.setVisible(true, true);
+		pt.setVisible(true);
+		// BM: not sure why, but this seems to be essential here to ensure that the ROI is visible
+		newPolyline.repaint(true, this.layer);
+		Display.repaint();
+		
+		// store directional data
+		this.gravDir = new RhizoGravitationalDirection(newPolyline, direction);		
+	}
+
+	/**
+	 * Get gravitational direction as angle in degrees.
+	 * @return	Angle of gravitational direction, 0 degrees point to the right.
+	 */
+	public double getGravitationalDirection() {
+		return this.gravDir.getDirection();
+	}
+
+	/**
+	 * Clear gravitational direction and remove polyline from layer.
+	 */
+	public void clearGravitationalDirection() {
+		
+    // find roi child of rootstack
+    ProjectThing dirProjectThing = this.getParentThing("gravitationaldirection");
+    if (dirProjectThing == null) {
+    	return;
+    }
+
+		// delete old direction if present
+		Set<Displayable> deleteSet = new HashSet<Displayable>();
+
+		for (ProjectThing ptc : dirProjectThing.findChildrenOfTypeR(Polyline.class)) {
+			Polyline pl = (Polyline) ptc.getObject();
+			if (layer == null || pl.getFirstLayer().getId() == layer.getId()) {
+				deleteSet.add(pl);
+			}
+		}
+
+		if (!deleteSet.isEmpty()) {
+			project.removeAll(deleteSet);
+		}
+	}
+	
+  /**
+   * Helper method to find parent thing in rhizoTrak project tree to keep given child.
+   * @param 	child		Identifier for parent thing.
+   * @return	Parent project thing.
+   */
+  private ProjectThing getParentThing(String child) {
+  	
+    ProjectTree projectTree = this.project.getProjectTree();
+    ProjectThing parentThing = RhizoUtils.getParentThingForChild(this.project, child);
+    if (parentThing == null) {
+
+    	// search for a rootstack object
+    	ProjectThing rootstack = RhizoUtils.getRootstacks(this.project).iterator().next();
+    	if (rootstack == null) {
+    		Utils.showMessage("rhizoTrak.setROI(): WARNING - cannot find a rootstack in project tree!");
+    		return null;
+    	}
+    	// add ROI things
+    	TemplateTree templateTree = this.project.getTemplateTree();
+    	TemplateThing template_root = this.project.getTemplateThing("rootstack");
+    	TemplateThing template_roi = templateTree.addNewChildType(template_root, child);
+    	templateTree.addNewChildType(template_roi, "polyline");
+
+    	// add ROI instance
+    	parentThing = rootstack.createChild(child);
+    	//add it to the tree
+			if (parentThing != null) {
+				DefaultMutableTreeNode parentNode = DNDTree.findNode(rootstack, projectTree);
+				DefaultMutableTreeNode new_node = new DefaultMutableTreeNode(parentThing);
+				((DefaultTreeModel)projectTree.getModel()).insertNodeInto(
+						new_node, parentNode, parentNode.getChildCount());
+				TreePath treePath = new TreePath(new_node.getPath());
+				projectTree.scrollPathToVisible(treePath);
+				projectTree.setSelectionPath(treePath);
+			}
+			// bring the display to front
+			if (parentThing.getObject() instanceof Displayable) {
+				Display.getFront().getFrame().toFront();
+			}
+    }
+    return parentThing;
+  }
 }